Move #include test to SourceKitTests and regenerate linux main
